Add support for the logging of selected events in Netconf.
A new startup parameter "-l" was added to netconf. The new parameter
specifies the name of a file that contains the logging options to be
enabled. The choices are:-
edit-config
get
get-config
kill-session
lock
unlock

The options are added in a single line separated by spaces e.g if
logging is only required when someone changes the config or kills
a session the logging option file would contain the line:-

edit-config kill-session.

The contents of the logging file can be modified at any time and Netconf
will dynamically use the new logging options.