Attempting to shutdown a netconf process can causes a crash
If a netconf session is active and a user attempts to shutdown
the netconf process, then a crash may occur. This happens most
frequently on very slow devices where the worker thread takes
some time to complete.

This patch has two changes:-
1) Improve the detection of the main_loop stopping in the worker
threads.
2) Making the termination handler shut down the worker threads and
wait for all threads to be complete before the process continues
with other shutdown routines.