Improve the handling of client hello messages
If a client sends a message with lots of capabilities set, then
potentially the fixed 4K server buffer would be insufficient to
receive the received message. This has been fixed by inspecting a
received hello message via recv with MSG_PEEK set. If a valid end
is found to the hello message then a buffer is dynamically allocated
to receive the message.