Fix with defaults - report-all failing some tests
The with default handling for netconf was mostly copied from the
equivalent apteryx-rest rest.c code. There are differences however
between the way restconf and netconf return data. Restconf only returns
information from the query end, but netconf returns a complete tree
from the top of the model. This means the way the with-defaults
is added needed to change.

The patch was authored by gcampbell512.